refactor: separate question-related endpoints into dedicated questions_router
