--- api/router/answer.py
@@ -1,33 +1,0 @@
-from fastapi import APIRouter, Depends
-from sqlmodel import Session, select
-
-from api.dependencies import get_assessment_id, get_session
-from api.schemas.answer import Answer, AnswerRead
-
-answer_router = APIRouter()
-
-
-@answer_router.get(
-    "/questions/{question_number}/answer",
-    tags=["answers"],
-    response_model=list[AnswerRead],
-    summary="Retrieve answer for question by user",
-    description="""
-Retrieve the latest user answers to the given question's tasks.
-""",
-)
-def get_answer(
-    question_number: int,
-    session: Session = Depends(get_session),
-    assessment_id: str = Depends(get_assessment_id),
-):
-    query = (
-        select(Answer)
-        .where(
-            Answer.question == question_number,
-            Answer.exam_id == assessment_id,
-            Answer.username == "hpotter",
-        )
-        .order_by(Answer.part, Answer.section, Answer.task)  # type: ignore
-    )
-    return session.exec(query).all()
